Add MCP integration for cost analysis
