+++ packages/server/shared/src/lib/system/system-prop.ts
@@ -107,6 +107,7 @@ export enum AppSystemProp {
   OPENOPS_MCP_SERVER_PATH = 'OPENOPS_MCP_SERVER_PATH',
 
   COST_EXPLORER_MCP_SERVER_PATH = 'COST_EXPLORER_MCP_SERVER_PATH',
+  COST_ANALYSIS_MCP_SERVER_PATH = 'COST_ANALYSIS_MCP_SERVER_PATH',
 }
 
 export enum SharedSystemProp {
